# Issue #44: Performance Optimization - Bundle Analysis & Runtime Monitoring