--- a/ISSUE_44_IMPLEMENTATION_PLAN.md
+++ b/ISSUE_44_IMPLEMENTATION_PLAN.md
@@ -1,0 +1,194 @@
+# Issue #44 Implementation Plan: Performance Optimization
+
+## Overview
+Optimize bundle size, runtime performance, and development experience for the modernized Swiss Army Knife Card with dual modern/legacy tool implementations.
+
+---
+
+## 🎯 Performance Optimization Areas
+
+### 1. Bundle Size Analysis & Optimization
+**Current Target**: <200KB gzipped
+**Priority**: HIGH
+
+#### Analysis Tasks
+- [ ] Generate bundle analysis report
+- [ ] Identify largest dependencies and modules
+- [ ] Analyze dual implementation overhead
+- [ ] Map unused code and dead code elimination opportunities
+
+#### Optimization Strategies
+- [ ] Implement code splitting for advanced visualization tools
+- [ ] Add lazy loading for less commonly used tools
+- [ ] Optimize import/export patterns
+- [ ] Tree-shaking optimization
+- [ ] Remove unused legacy code paths
+
+### 2. Runtime Performance Optimization
+**Target**: 60fps sustained, <100ms interaction response
+**Priority**: HIGH
+
+#### Performance Monitoring
+- [ ] Implement performance measurement hooks
+- [ ] Add rendering performance tracking
+- [ ] Monitor memory usage patterns
+- [ ] Track tool instantiation performance
+
+#### Optimization Implementation
+- [ ] Tool instance caching and reuse
+- [ ] Optimized re-rendering strategies
+- [ ] Efficient state update patterns
+- [ ] Memory leak prevention
+
+### 3. Development Performance
+**Target**: <30s build time, <5s dev server start
+**Priority**: MEDIUM
+
+#### Development Optimizations
+- [ ] Hot module replacement optimization
+- [ ] Development build time improvements
+- [ ] Faster test execution
+- [ ] Optimized linting and formatting
+
+---
+
+## 🔧 Implementation Strategy
+
+### Phase 1: Analysis & Measurement (4 hours)
+1. **Bundle Analysis** (2h) - Generate detailed bundle reports
+2. **Performance Baseline** (1h) - Establish current performance metrics
+3. **Optimization Planning** (1h) - Identify highest-impact optimizations
+
+### Phase 2: Bundle Optimization (6 hours)
+4. **Code Splitting** (3h) - Split advanced tools into separate chunks
+5. **Lazy Loading** (2h) - Implement dynamic imports for tools
+6. **Tree Shaking** (1h) - Optimize imports and exports
+
+### Phase 3: Runtime Optimization (4 hours)
+7. **Caching Implementation** (2h) - Tool instance and result caching
+8. **Rendering Optimization** (2h) - Efficient update patterns
+
+### Phase 4: Development Experience (2 hours)
+9. **Build Optimization** (1h) - Faster development builds
+10. **Monitoring Setup** (1h) - Performance monitoring infrastructure
+
+---
+
+## 📊 Performance Targets
+
+### Bundle Size Targets
+- **Total Bundle**: <200KB gzipped
+- **Core Bundle**: <150KB gzipped
+- **Advanced Tools**: <50KB gzipped (lazy loaded)
+- **Legacy Tools**: <100KB gzipped (on-demand)
+
+### Runtime Performance Targets
+- **Initial Load**: <2s on 3G connection
+- **Tool Instantiation**: <10ms per tool
+- **Re-render Performance**: <16ms (60fps)
+- **Memory Growth**: <1MB per hour
+- **Interaction Response**: <100ms
+
+### Development Targets
+- **Build Time**: <30s production, <10s development
+- **Dev Server Start**: <5s
+- **Test Execution**: <30s for full suite
+- **Hot Reload**: <1s for changes
+
+---
+
+## 🛠️ Optimization Techniques
+
+### Code Splitting Strategy
+```typescript
+// Dynamic imports for advanced tools
+const loadAdvancedTool = async (toolType: string) => {
+  switch (toolType) {
+    case 'gauge':
+      return (await import('./charts/GaugeTool.js')).GaugeTool;
+    case 'pie_chart':
+      return (await import('./charts/PieChartTool.js')).PieChartTool;
+    case 'heatmap':
+      return (await import('./charts/HeatmapTool.js')).HeatmapTool;
+    default:
+      return null;
+  }
+};
+```
+
+### Caching Implementation
+```typescript
+// Tool instance caching
+class ToolCache {
+  private static instances = new Map<string, BaseTool>();
+  
+  static get(key: string): BaseTool | undefined {
+    return this.instances.get(key);
+  }
+  
+  static set(key: string, tool: BaseTool): void {
+    this.instances.set(key, tool);
+  }
+}
+```
+
+### Performance Monitoring
+```typescript
+// Performance measurement hooks
+class PerformanceMonitor {
+  static measureRender(toolType: string, renderFn: () => any) {
+    const start = performance.now();
+    const result = renderFn();
+    const end = performance.now();
+    
+    if (end - start > 16) { // > 60fps threshold
+      console.warn(`Slow render: ${toolType} took ${end - start}ms`);
+    }
+    
+    return result;
+  }
+}
+```
+
+---
+
+## ✅ Success Criteria
+
+### Bundle Size Success
+- [ ] Total bundle <200KB gzipped
+- [ ] Core functionality <150KB
+- [ ] Advanced tools lazy loaded
+- [ ] No unused code in production bundle
+
+### Runtime Performance Success
+- [ ] 60fps sustained rendering
+- [ ] <100ms interaction response
+- [ ] Stable memory usage
+- [ ] Fast tool instantiation
+
+### Development Success
+- [ ] Faster build times
+- [ ] Hot module replacement working
+- [ ] Performance monitoring active
+- [ ] Developer tooling optimized
+
+---
+
+## 🎯 Expected Impact
+
+### Bundle Size Reduction
+- **Estimated Savings**: 20-30% through code splitting
+- **Lazy Loading**: 50KB+ saved on initial load
+- **Tree Shaking**: 10-15% reduction in unused code
+
+### Runtime Performance
+- **Rendering**: 15-20% improvement through caching
+- **Memory**: 25% reduction through efficient patterns
+- **Interactions**: 30% faster response times
+
+### Developer Experience
+- **Build Time**: 40-50% faster development builds
+- **Hot Reload**: Near-instant change reflection
+- **Debugging**: Better performance insights
+
+This optimization work ensures the modernized card performs excellently while maintaining the comprehensive feature set.