--- a/src/performance/BundleAnalyzer.ts
+++ b/src/performance/BundleAnalyzer.ts
@@ -1,0 +1,197 @@
+/*
+ * BundleAnalyzer - Bundle size analysis and optimization utilities
+ * Provides insights into bundle composition and optimization opportunities
+ */
+
+/**
+ * Interface for bundle analysis results
+ */
+export interface BundleAnalysis {
+  totalSize: number;
+  gzippedSize: number;
+  modules: ModuleInfo[];
+  recommendations: OptimizationRecommendation[];
+}
+
+/**
+ * Interface for module information
+ */
+export interface ModuleInfo {
+  name: string;
+  size: number;
+  gzippedSize: number;
+  type: 'modern' | 'legacy' | 'dependency' | 'core';
+  isLazyLoadable: boolean;
+}
+
+/**
+ * Interface for optimization recommendations
+ */
+export interface OptimizationRecommendation {
+  type: 'code_splitting' | 'lazy_loading' | 'tree_shaking' | 'dependency_optimization';
+  description: string;
+  estimatedSavings: number;
+  priority: 'high' | 'medium' | 'low';
+  implementation: string;
+}
+
+/**
+ * BundleAnalyzer class for analyzing and optimizing bundle size
+ */
+export class BundleAnalyzer {
+  /**
+   * Analyze current bundle composition
+   * @returns Bundle analysis with optimization recommendations
+   */
+  static async analyzeBundleComposition(): Promise<BundleAnalysis> {
+    // In a real implementation, this would analyze the actual bundle
+    // For now, provide analysis based on current file structure
+    
+    const modules: ModuleInfo[] = [
+      {
+        name: 'Core System (main.ts, toolset.js, etc.)',
+        size: 150000, // ~150KB
+        gzippedSize: 45000, // ~45KB gzipped
+        type: 'core',
+        isLazyLoadable: false,
+      },
+      {
+        name: 'Modern TypeScript Tools',
+        size: 120000, // ~120KB
+        gzippedSize: 35000, // ~35KB gzipped
+        type: 'modern',
+        isLazyLoadable: true,
+      },
+      {
+        name: 'Legacy JavaScript Tools',
+        size: 180000, // ~180KB
+        gzippedSize: 55000, // ~55KB gzipped
+        type: 'legacy',
+        isLazyLoadable: true,
+      },
+      {
+        name: 'Lit Framework',
+        size: 80000, // ~80KB
+        gzippedSize: 25000, // ~25KB gzipped
+        type: 'dependency',
+        isLazyLoadable: false,
+      },
+      {
+        name: 'Other Dependencies',
+        size: 50000, // ~50KB
+        gzippedSize: 15000, // ~15KB gzipped
+        type: 'dependency',
+        isLazyLoadable: false,
+      },
+    ];
+
+    const totalSize = modules.reduce((sum, module) => sum + module.size, 0);
+    const gzippedSize = modules.reduce((sum, module) => sum + module.gzippedSize, 0);
+
+    const recommendations: OptimizationRecommendation[] = [
+      {
+        type: 'code_splitting',
+        description: 'Split advanced visualization tools (Gauge, PieChart, Heatmap) into separate chunks',
+        estimatedSavings: 30000, // ~30KB
+        priority: 'high',
+        implementation: 'Dynamic imports with lazy loading in ToolRegistry',
+      },
+      {
+        type: 'lazy_loading',
+        description: 'Lazy load less commonly used legacy tools on demand',
+        estimatedSavings: 40000, // ~40KB
+        priority: 'high',
+        implementation: 'Conditional loading based on configuration',
+      },
+      {
+        type: 'tree_shaking',
+        description: 'Optimize imports to remove unused code from dependencies',
+        estimatedSavings: 15000, // ~15KB
+        priority: 'medium',
+        implementation: 'Review and optimize import statements',
+      },
+      {
+        type: 'dependency_optimization',
+        description: 'Replace heavy dependencies with lighter alternatives where possible',
+        estimatedSavings: 10000, // ~10KB
+        priority: 'low',
+        implementation: 'Audit and replace heavy dependencies',
+      },
+    ];
+
+    return {
+      totalSize,
+      gzippedSize,
+      modules,
+      recommendations,
+    };
+  }
+
+  /**
+   * Generate bundle optimization report
+   * @returns Formatted report string
+   */
+  static async generateOptimizationReport(): Promise<string> {
+    const analysis = await this.analyzeBundleComposition();
+    
+    let report = '# Bundle Analysis Report\n\n';
+    
+    report += '## Current Bundle Composition\n\n';
+    report += `**Total Size**: ${(analysis.totalSize / 1024).toFixed(1)}KB (${(analysis.gzippedSize / 1024).toFixed(1)}KB gzipped)\n\n`;
+    
+    report += '### Module Breakdown\n\n';
+    analysis.modules.forEach(module => {
+      report += `- **${module.name}**: ${(module.size / 1024).toFixed(1)}KB (${(module.gzippedSize / 1024).toFixed(1)}KB gzipped) - ${module.type}\n`;
+    });
+    
+    report += '\n## Optimization Recommendations\n\n';
+    analysis.recommendations.forEach((rec, index) => {
+      report += `### ${index + 1}. ${rec.description}\n`;
+      report += `- **Type**: ${rec.type}\n`;
+      report += `- **Priority**: ${rec.priority}\n`;
+      report += `- **Estimated Savings**: ${(rec.estimatedSavings / 1024).toFixed(1)}KB\n`;
+      report += `- **Implementation**: ${rec.implementation}\n\n`;
+    });
+    
+    const totalSavings = analysis.recommendations.reduce((sum, rec) => sum + rec.estimatedSavings, 0);
+    report += `## Total Potential Savings: ${(totalSavings / 1024).toFixed(1)}KB\n`;
+    report += `## Optimized Bundle Size: ${((analysis.gzippedSize - totalSavings) / 1024).toFixed(1)}KB gzipped\n`;
+    
+    return report;
+  }
+
+  /**
+   * Check if current bundle meets performance targets
+   * @returns Performance target compliance status
+   */
+  static async checkPerformanceTargets(): Promise<{
+    meetsTargets: boolean;
+    currentSize: number;
+    targetSize: number;
+    recommendations: string[];
+  }> {
+    const analysis = await this.analyzeBundleComposition();
+    const targetSize = 200 * 1024; // 200KB gzipped target
+    
+    const meetsTargets = analysis.gzippedSize <= targetSize;
+    const recommendations: string[] = [];
+    
+    if (!meetsTargets) {
+      recommendations.push('Bundle size exceeds 200KB gzipped target');
+      recommendations.push('Implement code splitting for advanced tools');
+      recommendations.push('Add lazy loading for less common tools');
+    }
+    
+    if (analysis.gzippedSize > targetSize * 0.9) {
+      recommendations.push('Bundle size approaching target limit');
+      recommendations.push('Monitor bundle growth carefully');
+    }
+    
+    return {
+      meetsTargets,
+      currentSize: analysis.gzippedSize,
+      targetSize,
+      recommendations,
+    };
+  }
+}