… docs/v3/PHASE5_CORE_MODERNIZATION_PLAN.md